Prompt user prior to deleting memory records, added exception handling

+++ src/brain.py
@@ -2,13 +2,12 @@ from fuzzywuzzy import fuzz
 import datetime
 import os
 
-from . import io, user, data, memory, ui
+from . import io, user, data, memory
 
 class Brain:
     def __init__(self, mem_db_path):
         self.mem_db = io.DB(mem_db_path)
         self.memories = [memory.Memory(key,t,k,b) for key,t,k,b in self.mem_db.dump()]
-
         self.top_n = 10
 
         # words to omit from fuzzy string search, e.g. and the is are etc.
@@ -40,7 +39,7 @@ class Brain:
     def create_memory(self):
         try:
             # display UI so user can fill in memory data
-            mem_ui = ui.MemDisplay()
+            mem_ui = memory.UI()
             self.mem_db.insert(mem_ui.title_text, mem_ui.keywords_text, mem_ui.body_text)
         except KeyboardInterrupt:
             print('Discarded new memory data')
@@ -54,8 +53,11 @@ class Brain:
         while True:
             Mem = self._select_memory_from_list(m_matches, 'Display')
             if Mem:
-                mem_ui = ui.MemDisplay(Mem.title, Mem.keywords, Mem.body)
-                self.mem_db.update(Mem, mem_ui)
+                try:
+                    mem_ui = memory.UI(Mem.title, Mem.keywords, Mem.body)
+                    self.mem_db.update(Mem, mem_ui)
+                except KeyboardInterrupt:
+                    pass
             else:
                 break
 
@@ -91,7 +93,7 @@ class Brain:
         m_matches = self._memory_match(user_keywords)
         while True:
             Mem = self._select_memory_from_list(m_matches, 'Remove')
-            if Mem:
+            if Mem and user.confirm('delete'):
                 self.mem_db.remove(Mem.db_key)
                 self.memories.remove(Mem)
                 m_matches.remove(Mem)
